refactor: shared form validators and state matcher
