+++ libs/shared/src/modules/forms/errorStateMatcher.ts
@@ -1,0 +1,13 @@
+import { FormControl, FormGroupDirective, NgForm } from '@angular/forms';
+import { ErrorStateMatcher } from '@angular/material/core';
+
+/**
+ * Implement custom error handler to only display if control is dirty, touched, or submitted.
+ * https://material.angular.io/components/input/overview#changing-when-error-messages-are-shown
+ * */
+export class showErrorAfterInteraction implements ErrorStateMatcher {
+  isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
+    const isSubmitted = form && form.submitted;
+    return !!(control && control.invalid && (control.dirty || control.touched || isSubmitted));
+  }
+}
