+++ libs/shared/src/modules/forms/index.ts
@@ -1,0 +1,3 @@
+export * from './components';
+export * from './forms.module';
+export * from './validators';
