+++ libs/shared/src/modules/forms/validators/index.ts
@@ -1,0 +1,7 @@
+import isUrl from './isUrl';
+import passwordMatch from './passwordMatch';
+
+export const PICSAFormValidators = {
+  isUrl,
+  passwordMatch,
+};
